moved the upload to its own file
